+++ Proyecto2.5/Assets/scripts/DestroyPlatform.cs
@@ -1,16 +1,6 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
-
 public class DestroyPlatform : MonoBehaviour
 {
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
     // Update is called once per frame
     private void OnTriggerEnter2D(Collider2D other)
     {
