+++ Proyecto2.5/Assets/scripts/DestroyPlatform.cs.meta
