+++ Proyecto2.5/Assets/scripts/Disparo.cs.meta
