+++ Proyecto2.5/Assets/scripts/Movimientobala.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+public class Movimientobala : MonoBehaviour
+{
+    [SerializeField] private float velocidad=4f;
+    void Update()
+    {
+        transform.Translate(Vector3.up*velocidad*Time.deltaTime);  
+    }
+}
