+++ Proyecto2.5/Assets/scripts/SceneManagers.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+using UnityEngine.SceneManagement;
+public class SceneManagers : MonoBehaviour
+{
+    public void RestartScene()
+    {
+        SceneManager.LoadScene(0);
+    }
+}
