+++ Proyecto2.5/Assets/scripts/morir.cs
@@ -1,23 +1,8 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.SceneManagement;
 public class morir : MonoBehaviour
 {
     [SerializeField] private GameObject adios;
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-          
-    }
-
     private void OnCollisionEnter2D(Collision2D other)
     {
         if (other.gameObject.CompareTag("Platform"))
@@ -26,6 +11,5 @@ public class morir : MonoBehaviour
             adios.SetActive(true);
         }
     }
-    
 }
