+++ Proyecto2.5/Assets/scripts/morir.cs.meta
