+++ Proyecto2.5/ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 2021.3.8f1
-m_EditorVersionWithRevision: 2021.3.8f1 (b30333d56e81)
+m_EditorVersion: 2021.3.10f1
+m_EditorVersionWithRevision: 2021.3.10f1 (1c7d0df0160b)
